Ignore udev's FHS violating large binary cache file /etc/udev/hwdb.bin
On debian it is not present by default there but in /lib/udev/hwdb.bin,
but I think it gets created if /etc/udev/hwdb.d/ has files added to it;
see hwdb(7).

This file is a FHS violation because /etc is for configuration. It boggles
my mind that udev willfully violates the FHS. Shame. Shame. Shame.